services/doc-processor/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
python-dotenv==1.0.0
